Ignores: method 3 - use custom File

+++ src/bootstrap-outdated.php
@@ -1,8 +1,8 @@
 <?php declare(strict_types=1);
 
-$outdatedVirtualFileOriginal = tempnam(sys_get_temp_dir(), 'phpcs-ignores-outdated');
-$outdatedVirtualFile = $outdatedVirtualFileOriginal . '.php';
-file_put_contents($outdatedVirtualFile, '<?php');
-@unlink($outdatedVirtualFileOriginal); // intentionally @ - file may not exists
+$outdatedVirtualFile = sys_get_temp_dir() . '/phpcs-ignores-outdated.php';
+if (!file_exists($outdatedVirtualFile)) {
+	file_put_contents($outdatedVirtualFile, '<?php');
+}
 
 require_once __DIR__ . '/bootstrap.php';
